requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic[email]==2.5.0
email-validator==2.1.0
python-multipart==0.0.6
httpx==0.25.2
python-dotenv==1.0.0
jinja2==3.1.2

# Data processing
pandas==2.1.4
numpy==1.26.2

# AI/ML dependencies
openai==1.12.0
chromadb==0.4.22
sentence-transformers==2.3.1
beautifulsoup4==4.12.3
lxml==5.1.0

# Optional dependencies (only if needed)
# psycopg2-binary==2.9.9
# pgvector==0.2.4
# asyncpg==0.29.0
# celery==5.3.4
# redis==5.0.1
# alembic==1.12.1